requirements.txt
# Base ----------------------------------------
sqlalchemy
pymysql
pandas
python-dotenv
psycopg2-binary
requests
spotipy

numpy>=1.18.5
opencv-python>=4.1.2
matplotlib>=3.2.2
ipyleaflet>=0.14.0
sympy>=1.10.1

# Machine learning ----------------------------
scikit-learn